Move the color harmony function to the component that will use it, until we have something else to use it on
