+++ static/data/hubspot.json
@@ -232,5 +232,5 @@
   "totalDonations": 174,
   "totalMonthlyDonations": 102,
   "totalOneTimeDonations": 72,
-  "lastUpdated": "2025-03-10T20:28:28.041Z"
+  "lastUpdated": "2025-03-10T23:07:03.225Z"
 }
